CLOUDP-62721: Move Atlas E2E to an evergreen task
