+++ .evergreen.yml
@@ -41,6 +41,30 @@ functions:
           GOROOT: ${go_root}
           <<: *go_options
         command: make test
+  "atlas e2e":
+    - command: subprocess.exec
+        type: test
+        params:
+          add_to_path:
+            - ${go_bin}
+          working_dir: src/github.com/mongodb/mongocli/e2e/atlas
+          env:
+            WORKDIR: ${workdir}
+            GOROOT: ${go_root}
+            <<: *go_options
+          command: make e2e-test-atlas
+  "iam e2e":
+    - command: subprocess.exec
+        type: test
+        params:
+          add_to_path:
+            - ${go_bin}
+          working_dir: src/github.com/mongodb/mongocli/e2e/iam
+          env:
+            WORKDIR: ${workdir}
+            GOROOT: ${go_root}
+            <<: *go_options
+          command: make e2e-test-iam
   "lint":
     - command: subprocess.exec
       type: test
@@ -292,6 +316,24 @@ tasks:
     commands:
       - func: "clone"
       - func: "test"
+  - name: atlas_e2e
+    command: "atlas e2e"
+    params:
+      MCLI_OPS_MANAGER_URL: ${mcli_ops_manager_url}
+      MCLI_ORG_ID: ${mcli_org_id}
+      MCLI_PRIVATE_API_KEY: ${mcli_private_api_key}
+      MCLI_PROJECT_ID: ${mcli_project_id}
+      MCLI_PUBLIC_API_KEY: ${mcli_public_api_key}
+      MCLI_SERVICE: ${mcli_service}
+  - name: iam_e2e
+    command: "iam e2e"
+    params:
+      MCLI_OPS_MANAGER_URL: ${mcli_ops_manager_url}
+      MCLI_ORG_ID: ${mcli_org_id}
+      MCLI_PRIVATE_API_KEY: ${mcli_private_api_key}
+      MCLI_PROJECT_ID: ${mcli_project_id}
+      MCLI_PUBLIC_API_KEY: ${mcli_public_api_key}
+      MCLI_SERVICE: ${mcli_service}
   - name: package_msi
     tags: ["packaging"]
     depends_on:
@@ -594,6 +636,17 @@ buildvariants:
     tasks:
       - name: compile
       - name: test
+  - name: e2e
+    display_name: "E2E"
+    run_on:
+      - rhel70-small
+    expansions:
+      go_root: "/opt/golang/go1.14"
+      go_bin: "/opt/golang/go1.14/bin"
+    tasks:
+      - name: compile
+      - name: atlas_e2e
+      - name: iam_e2e
   - name: goreleaser_snaphot
     display_name: "Packaging and Publishing (goreleaser, apt, yum)"
     run_on:
