+++ Makefile
@@ -90,6 +90,18 @@ build: ## Generate a binary in ./bin
 	@echo "==> Building binary"
 	go build -ldflags "${LINKER_FLAGS}" -o ${DESTINATION}
 
+.PHONY: e2e-test-atlas
+e2e-test-atlas: build ## Run Atlas E2E tests
+	@echo "==> Running E2E tests..."
+	# the target assumes the MCLI-* environment variables are exported
+	go test -v -p 1 -parallel 1 -tags=e2e ./e2e/atlas...
+
+.PHONY: e2e-test-iam
+e2e-test-iam: build ## Run IAM E2E tests
+	@echo "==> Running E2E tests..."
+	# the target assumes the MCLI-* environment variables are exported
+	go test -v -p 1 -parallel 1 -tags=e2e ./e2e/iam...
+
 .PHONY: e2e-test
 e2e-test: build ## Run E2E tests
 	@echo "==> Running E2E tests..."
